\name{run_landclim}
\alias{run_landclim}

\title{
Run LandClim
}
\description{

}
\usage{
run_landclim(control_file)
}

\arguments{
  \item{control_file}{Name of the control file (e.g. control.xml).

}
}
\details{
The working directory needs to be set to the level above "Input" and "Output" folders of LandClim, thus at the "site" level. Until now, the folder names are hard-coded, thus need to be "Input" and "Output", not e.g. input! Feel free to fix this!

You need to set the path to your LandClim executable before you run LandClim using this function.

}

\references{
### Add LandClim refrences
}
\author{
Klara Dolos
}

\seealso{
set_landclim_path
}
\examples{

}


